requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
diskcache==5.6.3
docker==7.1.0
docutils==0.21.2
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.25
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
nh3==0.2.21
nltk==3.9.1
numpy==2.2.6
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pyarrow==20.0.0
pyautogen==0.9.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
twine==6.1.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xgboost==3.0.2
zstandard==0.23.0
